Don't pass unencrypted seedPhrase around
It's probably better that we just decrypt the seed phrase and pass it straight into market maker, then let it get garbage collected. If we keep it attached on our portfolio reference there's a chance it'll get accidentily written in plain text to a log file or dumped on the screen in an error dialog.