Store currencies in main App state
This gives us a single source of truth to refer to application wide. If we use `props.currencies` throughout the app it means we can be confident we're always using the correct up to date balance and are always showing the currently activated coins.

It also means we avoid fetching/storing/updating currency activation/balance state in all the components that require it and keeping them in sync with eachother.